% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARCD.R
\name{ARCD}
\alias{ARCD}
\title{ARCD, Data Dictionary Archiver}
\description{
ARCD class and its methods
}
\examples{
\donttest{
e <- new.env()
# this would work if you had these REDCap projects, API access, and shelter
unlockREDCap(c('rdb_users', 'rdb_data'), 'https://redcap.vumc.org/api/', 'redcapAPI', envir = e)
x <- ARCD$new()
# `pull` and `set` REDCap metadata
md <- pull(e[['rdb_data']])
x$set(md)
# or do it in a single step
x$set(e[['rdb_users']])

# show contents of archive
x$ls()
# retrieve an archived data dictionary
dd <- x$get(x$ls()$tname[1])
dim(dd)
# show location of "duckdb" database
x

t1 <- x$ls()$tname[1]
t2 <- x$ls()$tname[2]
# "simple" difference
x$sdiff(t1, t2)
# show difference in web browser
x$diff(t1, t2)
}
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{con}}{\link{DBIConnection}\cr
\sQuote{duckdb} database connection.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ARCD-new}{\code{ARCD$new()}}
\item \href{#method-ARCD-print}{\code{ARCD$print()}}
\item \href{#method-ARCD-setup}{\code{ARCD$setup()}}
\item \href{#method-ARCD-ls}{\code{ARCD$ls()}}
\item \href{#method-ARCD-get}{\code{ARCD$get()}}
\item \href{#method-ARCD-set}{\code{ARCD$set()}}
\item \href{#method-ARCD-sdiff}{\code{ARCD$sdiff()}}
\item \href{#method-ARCD-diff}{\code{ARCD$diff()}}
\item \href{#method-ARCD-destroy}{\code{ARCD$destroy()}}
\item \href{#method-ARCD-clone}{\code{ARCD$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARCD-new"></a>}}
\if{latex}{\out{\hypertarget{method-ARCD-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} ARCD class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARCD$new(dbdir = "ARCD.duckdb")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbdir}}{(\code{character(1)})\cr
A file name for a duckdb database; defaults to "ARCD.duckdb".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARCD-print"></a>}}
\if{latex}{\out{\hypertarget{method-ARCD-print}{}}}
\subsection{Method \code{print()}}{
Print method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARCD$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to \code{cat}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
itself invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARCD-setup"></a>}}
\if{latex}{\out{\hypertarget{method-ARCD-setup}{}}}
\subsection{Method \code{setup()}}{
Setup method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARCD$setup()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
itself invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARCD-ls"></a>}}
\if{latex}{\out{\hypertarget{method-ARCD-ls}{}}}
\subsection{Method \code{ls()}}{
List all archived data dictionaries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARCD$ls()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\sQuote{dbGetQuery} result
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARCD-get"></a>}}
\if{latex}{\out{\hypertarget{method-ARCD-get}{}}}
\subsection{Method \code{get()}}{
Retrieve an archived data dictionary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARCD$get(tname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tname}}{(\code{character(1)}) table name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\sQuote{dbGetQuery} result.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARCD-set"></a>}}
\if{latex}{\out{\hypertarget{method-ARCD-set}{}}}
\subsection{Method \code{set()}}{
Show simple difference between two data dictionaries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARCD$set(md)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{md}}{metadata or a \sQuote{redcapConnection}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE for success or FALSE for failure.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARCD-sdiff"></a>}}
\if{latex}{\out{\hypertarget{method-ARCD-sdiff}{}}}
\subsection{Method \code{sdiff()}}{
Show simple difference between two data dictionaries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARCD$sdiff(tnameA, tnameB)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tnameA}}{(\code{character(1)}) first table name.}

\item{\code{tnameB}}{(\code{character(1)}) second table name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
numeric vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARCD-diff"></a>}}
\if{latex}{\out{\hypertarget{method-ARCD-diff}{}}}
\subsection{Method \code{diff()}}{
Show difference between two data dictionaries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARCD$diff(tnameA, tnameB, file_name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tnameA}}{(\code{character(1)}) first table name.}

\item{\code{tnameB}}{(\code{character(1)}) second table name.}

\item{\code{file_name}}{(\code{character(1)}) optional file name to save differences as HTML.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
HTML output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARCD-destroy"></a>}}
\if{latex}{\out{\hypertarget{method-ARCD-destroy}{}}}
\subsection{Method \code{destroy()}}{
Delete the duckdb database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARCD$destroy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ARCD-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ARCD-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ARCD$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
