% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivedd-package.R
\docType{package}
\name{archivedd-package}
\alias{archivedd}
\alias{archivedd-package}
\title{Archive Data Dictionary}
\description{
It archives and compares REDCap data dictionaries.
}
\details{
This package is experimental. Please submit bugs to
\url{https://github.com/couthcommander/archivedd}.
}
\examples{
\donttest{
e <- new.env()
# this would work if you had these REDCap projects, API access, and shelter
unlockREDCap(c('rdb_users', 'rdb_data'), 'https://redcap.vumc.org/api/', 'redcapAPI', envir = e)
x <- ARCD$new()
x$set(e[['rdb_data']])
x$set(e[['rdb_users']])
t1 <- x$ls()$tname[1]
t2 <- x$ls()$tname[2]
x$diff(t1, t2)
}
}
\author{
Cole Beck \email{cole.beck@vumc.org}

Maintainer: Cole Beck \email{cole.beck@vumc.org}
}
