% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull}
\alias{pull}
\title{Pull REDCap Data Dictionary}
\usage{
pull(rcon)
}
\arguments{
\item{rcon}{A redcapConnection object.}
}
\description{
Print method for \sQuote{packageInfo} class.
}
\details{
Print its argument and return it invisibly.
}
\examples{
\donttest{
e <- new.env()
# this would work if you had these REDCap projects, API access, and shelter
unlockREDCap(c('rdb_users', 'rdb_data'), 'https://redcap.vumc.org/api/', 'redcapAPI', envir = e)
md <- pull(e[['rdb_data']])
md$info
dim(md$data)
}
}
